Remind people to sign their names
ActionNetwork is *supposed* to prepend a salutation to the recipient of the letter campaign, as well as append the activist's sign off... but from what I can tell, it's not actually doing it.

So I added a reminder for them to do this.

I didn't want to leave placeholder text in the pre-written letter because if they forget to do it, it looks worse than just having an empty ending in my opinion.